Working on first part of user story
I want to get the first part of user story done
which is have the app ask what they want to log today.
To do that we need to be able to/already have a way to track
calories for the day.  Once we can do that then we can ask the user
to add an item and save it and keep a total.  In the future
will add capability to add breakfast lunch and dinner
which will be there own lists.  Today the goal is let the user
add an item. And keep an accurate running total of calories for the day.
and be able to save multiple days. And have the program know what
day it is and if the user has added anything for it.